# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kahflane API",
    "description": "AI-powered Knowledge Management System with Multi-Tenant Architecture",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/tenants": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Get User Tenants",
        "description": "Get all tenants the current user has access to.",
        "operationId": "get_user_tenants_api_v1_auth_tenants_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserTenantsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/tenants/{slug}": {
      "get": {
        "tags": [
          "tenants"
        ],
        "summary": "Get Tenant",
        "description": "Get tenant information by slug.\n\nThis endpoint is public to allow checking tenant existence.",
        "operationId": "get_tenant_api_v1_tenants__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenants/": {
      "get": {
        "tags": [
          "tenants"
        ],
        "summary": "Get Current Tenant Info",
        "description": "Get current tenant information.\n\nRequires authentication.",
        "operationId": "get_current_tenant_info_api_v1_tenants__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "tenants"
        ],
        "summary": "Update Tenant",
        "description": "Update current tenant settings.\n\nOnly tenant owners can update settings.",
        "operationId": "update_tenant_api_v1_tenants__patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/tenants/invite": {
      "post": {
        "tags": [
          "tenants"
        ],
        "summary": "Invite Member",
        "description": "Invite a user to join the current tenant. Requires admin or owner role.",
        "operationId": "invite_member_api_v1_tenants_invite_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/tenants/invitations/{token}/accept": {
      "post": {
        "tags": [
          "tenants"
        ],
        "summary": "Accept Invitation",
        "description": "Accept a tenant invitation. If the user doesn't exist, password and\nfull_name are required to create an account.",
        "operationId": "accept_invitation_api_v1_tenants_invitations__token__accept_post",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenants/leave": {
      "post": {
        "tags": [
          "tenants"
        ],
        "summary": "Leave Tenant",
        "description": "Leave the current tenant. Owners cannot leave if they are the last owner.",
        "operationId": "leave_tenant_api_v1_tenants_leave_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/tenants/default": {
      "put": {
        "tags": [
          "tenants"
        ],
        "summary": "Switch Default Tenant",
        "description": "Switch the user's default tenant.",
        "operationId": "switch_default_tenant_api_v1_tenants_default_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchDefaultTenantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcceptInvitationRequest": {
        "properties": {
          "password": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 8
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          }
        },
        "type": "object",
        "title": "AcceptInvitationRequest",
        "description": "Request to accept a tenant invitation."
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "examples": [
              "Resource not found"
            ]
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "Standard error response."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InvitationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "token": {
            "type": "string",
            "title": "Token"
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "role",
          "token",
          "expires_at",
          "message"
        ],
        "title": "InvitationResponse",
        "description": "Response for created invitation."
      },
      "InviteMemberRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "examples": [
              "jane@example.com"
            ]
          },
          "role": {
            "type": "string",
            "pattern": "^(admin|member)$",
            "title": "Role",
            "default": "member",
            "examples": [
              "member"
            ]
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "InviteMemberRequest",
        "description": "Request to invite a member to the tenant."
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "description": "Generic message response."
      },
...
      "SwitchDefaultTenantRequest": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id",
            "examples": [
              "7c9e6679-7425-40de-944b-e07fc1f90ae7"
            ]
          }
        },
        "type": "object",
        "required": [
          "tenant_id"
        ],
        "title": "SwitchDefaultTenantRequest",
        "description": "Request to switch default tenant."
      },
...
      "TenantInfo": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id",
            "examples": [
              "7c9e6679-7425-40de-944b-e07fc1f90ae7"
            ]
          },
          "slug": {
            "type": "string",
            "title": "Slug",
            "examples": [
              "acme-corp"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Acme Corp"
            ]
          },
          "role": {
            "type": "string",
            "title": "Role",
            "examples": [
              "owner"
            ]
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default",
            "examples": [
              true
            ]
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "slug",
          "name",
          "role",
          "is_default"
        ],
        "title": "TenantInfo",
        "description": "Tenant information for user."
      },
      "TenantResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "examples": [
              "7c9e6679-7425-40de-944b-e07fc1f90ae7"
            ]
          },
          "slug": {
            "type": "string",
            "title": "Slug",
            "examples": [
              "acme-corp"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Acme Corp"
            ]
          },
          "domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain",
            "examples": [
              "acme.com"
            ]
          },
          "plan_type": {
            "type": "string",
            "title": "Plan Type",
            "examples": [
              "free"
            ]
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "examples": [
              true
            ]
          },
          "settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Settings",
            "examples": [
              {
                "feature_flags": {
                  "ai_search": true
                }
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "plan_type",
          "is_active",
          "settings"
        ],
        "title": "TenantResponse",
        "description": "Tenant information response."
      },
...
      "UpdateTenantRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "examples": [
              "Acme Corporation"
            ]
          },
          "domain": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain",
            "examples": [
              "acme.com"
            ]
          },
          "settings": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Settings",
            "examples": [
              {
                "feature_flags": {
                  "ai_search": true
                }
              }
            ]
          }
        },
        "type": "object",
        "title": "UpdateTenantRequest",
        "description": "Request to update tenant settings."
      },
...
      "UserTenantsResponse": {
        "properties": {
          "tenants": {
            "items": {
              "$ref": "#/components/schemas/TenantInfo"
            },
            "type": "array",
            "title": "Tenants"
          }
        },
        "type": "object",
        "required": [
          "tenants"
        ],
        "title": "UserTenantsResponse",
        "description": "Response listing user's tenants."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}